Parallax se mueve con lerp, cámara aun no, y spritesheets de canelon

+++ Project Stroke/Proyecto/ecs/Camera.h
@@ -22,7 +22,7 @@ private:
 	SDL_Rect camera_;
 	Vector2D camPos;
 	int players;
-	int heightMap_= 0;
+	float heightMap_= 0;
 
 	void followPlayer();
 	void followBossCat();
